RAG/part2/requirements.txt
google-generativeai
python-dotenv
pandas
requests
numpy
groq
sentence-transformers